+++ tine20/Tinebase/Config.php
@@ -1375,7 +1375,6 @@ class Tinebase_Config extends Tinebase_Config_Abstract
          * backend (string) - see Tinebase_EmailUser::$_supportedBackends
          * verifyPeer (bool)
          * "allowOverwrite": false (bool)
-         * "allowExternalEmail": false (bool) // deprecated, please use smtp.allowAnyExternalDomains instead
          *
          * TODO make this a structured config with subconfig keys
          */